+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 * Detect necessity reload `aggregator` model if it was `detail` model modified.
+* Transition to edit form route when record has an `id` and same time is new.
 
 ## [0.7.0-beta.13] - 2016-10-07
 ### Changed